requirements.txt
torch >= 2.11.0
numpy
Pillow